+++ generators/src/lib/engine/index.ts
@@ -1,3 +1,3 @@
 export * from './generator-utils'
 export * from './generator-types'
-export * from './crud-auth'
+export * from './crud-security'